+++ template/theme-support/twentynineteen/content-single.php
@@ -1,0 +1,29 @@
+<article id="post-<?php the_ID(); ?>" <?php post_class(); ?>>
+	<header class="entry-header">
+		<?php get_template_part( 'template-parts/header/entry', 'header' ); ?>
+	</header>
+
+	<?php twentynineteen_post_thumbnail(); ?>
+
+	<div class="entry-content">
+		<?php
+		the_content();
+
+		wp_link_pages(
+			array(
+				'before' => '<div class="page-links">' . __( 'Pages:', 'twentynineteen' ),
+				'after'  => '</div>',
+			)
+		);
+		?>
+	</div><!-- .entry-content -->
+
+	<footer class="entry-footer">
+		<?php twentynineteen_entry_footer(); ?>
+	</footer><!-- .entry-footer -->
+
+	<?php if ( ! is_singular( 'attachment' ) ) : ?>
+	<?php get_template_part( 'template-parts/post/author', 'bio' ); ?>
+	<?php endif; ?>
+
+</article><!-- #post-${ID} -->